Change handling of GOOGLE_CLOUD_PROJECT in build.sh.
The previous behavior could possibly, conceivably be causing problems.
Now, the user now specifies GOOGLE_CLOUD_PROJECT_FOR_TESTS, which
should not collide with anything, and the GOOGLE_CLOUD_PROJECT
environment variable is only set for the single thing that should need
it, which is run_system_tests.sh, and only in the single, ephemeral
Docker container where we run those system tests.