Revert to previous virtualenv behavior.
A previous change caused the 'virtualenv' command change to switch its
default Python version from 2.7 to 3.6 (when no -p flag is passed).
Although not a bad change, it was accidental, and this reverts to the
previous behavior.  This also add tests for the behavior, and fixes
various tests that had been subtly broken by a previous refactoring.

The cause of this behavior change was that the 'virtualenv' package
was installed under each Python interpreter, instead of the 'wheel'
package.